requirements_prod.txt
colorama
numpy==1.23.5
pandas==1.5.3
pygeohash
scikit-learn==1.3.1
tensorflow==2.10.0
google-cloud-bigquery
google-cloud-storage==2.14.0
google-api-core==2.8.2
googleapis-common-protos==1.56.4
protobuf==3.19.6
h5py==3.10.0
pyarrow
typing_extensions==4.8.0
mlflow==2.1.1
fastapi==0.104.1
pydantic==2.4.2
pytz
uvicorn
python-dotenv


openai==1.54.4
